refactor: Move VoyageurApp to another file
previously VoyageurApp was in MainActiviy.kt. Moved it to App.kt
