+++ app/src/main/java/com/android/voyageur/App.kt
@@ -1,0 +1,40 @@
+package com.android.voyageur
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.navigation
+import androidx.navigation.compose.rememberNavController
+import com.android.voyageur.ui.navigation.NavigationActions
+import com.android.voyageur.ui.navigation.Route
+import com.android.voyageur.ui.navigation.Screen
+import com.android.voyageur.ui.overview.OverviewScreen
+import com.android.voyageur.ui.profile.ProfileScreen
+import com.android.voyageur.ui.search.SearchScreen
+
+@Composable
+fun VoyageurApp() {
+  val navController = rememberNavController()
+  val navigationActions = NavigationActions(navController)
+
+  NavHost(navController = navController, startDestination = Route.OVERVIEW) {
+    navigation(
+        startDestination = Screen.OVERVIEW,
+        route = Route.OVERVIEW,
+    ) {
+      composable(Screen.OVERVIEW) { OverviewScreen(navigationActions) }
+    }
+    navigation(
+        startDestination = Screen.SEARCH,
+        route = Route.SEARCH,
+    ) {
+      composable(Screen.SEARCH) { SearchScreen(navigationActions) }
+    }
+    navigation(
+        startDestination = Screen.PROFILE,
+        route = Route.PROFILE,
+    ) {
+      composable(Screen.PROFILE) { ProfileScreen(navigationActions) }
+    }
+  }
+}
