+++ app/src/main/java/com/android/voyageur/MainActivity.kt
@@ -6,21 +6,10 @@ import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTag
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.navigation
-import androidx.navigation.compose.rememberNavController
 import com.android.voyageur.resources.C
-import com.android.voyageur.ui.navigation.NavigationActions
-import com.android.voyageur.ui.navigation.Route
-import com.android.voyageur.ui.navigation.Screen
-import com.android.voyageur.ui.overview.OverviewScreen
-import com.android.voyageur.ui.profile.ProfileScreen
-import com.android.voyageur.ui.search.SearchScreen
 import com.android.voyageur.ui.theme.SampleAppTheme
 
 class MainActivity : ComponentActivity() {
@@ -35,33 +24,6 @@ class MainActivity : ComponentActivity() {
               VoyageurApp()
             }
       }
-    }
-  }
-}
-
-@Composable
-fun VoyageurApp() {
-  val navController = rememberNavController()
-  val navigationActions = NavigationActions(navController)
-
-  NavHost(navController = navController, startDestination = Route.OVERVIEW) {
-    navigation(
-        startDestination = Screen.OVERVIEW,
-        route = Route.OVERVIEW,
-    ) {
-      composable(Screen.OVERVIEW) { OverviewScreen(navigationActions) }
-    }
-    navigation(
-        startDestination = Screen.SEARCH,
-        route = Route.SEARCH,
-    ) {
-      composable(Screen.SEARCH) { SearchScreen(navigationActions) }
-    }
-    navigation(
-        startDestination = Screen.PROFILE,
-        route = Route.PROFILE,
-    ) {
-      composable(Screen.PROFILE) { ProfileScreen(navigationActions) }
     }
   }
 }
